cleanup + fix seq note off
