+++ Source/AliveLibCommon/audio/Stream.hpp
@@ -7,6 +7,23 @@
 
 namespace psx {
 
+class Exception : public std::exception
+{
+public:
+    explicit Exception(const char* msg)
+        : mMsg(msg)
+    {
+    }
+
+    const char* what() const throw() override
+    {
+        return mMsg;
+    }
+
+private:
+    const char* mMsg;
+};
+
 class Stream
 {
 public:
